Final Commit
Ready for Review. Added in a video with the sources of mp4 and webm. Added in a poster for that video. Made a captions file for the video. Added in those caption to the video. Styles the video to be floated on the right when the screen is big enough. Added in styles for the captions that set the fontsize, backgroundcolor, and fontcolor. Renamed all of the files to their proper names. Added in comments to the file.